backend/requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-base64==2.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==6.1.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
Pillow==9.2.0
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
